requirements.txt
streamlit>=1.29.0
google-genai>=0.2.0
pandas>=2.2.0
altair>=5.2.0
python-dotenv>=1.0.0
